// gen/resources_pb.ts
export class Resource extends Message<Resource> {
  /**
   * The resources's ID.
   *
   * This uniquely identifies this resource across all *currently live*
   * resources. This is also the primary way any operations on a resource
   * are associated with it
   *
   * @generated from field: rs.tokio.console.common.Id id = 1;
   */
  id?: Id;

  /**
   * The numeric ID of the resources's `Metadata`.
   *
   * @generated from field: rs.tokio.console.common.MetaId metadata = 2;
   */
  metadata?: MetaId;

  /**
   * The resources's concrete rust type.
   *
   * @generated from field: string concrete_type = 3;
   */
  concreteType = "";

  /**
   * The kind of resource (e.g timer, mutex)
   *
   * @generated from field: rs.tokio.console.resources.Resource.Kind kind = 4;
   */
  kind?: Resource_Kind;

  /**
   * The location in code where the resource was created.
   *
   * @generated from field: rs.tokio.console.common.Location location = 5;
   */
  location?: Location;

  /**
   * The ID of the parent resource.
   *
   * @generated from field: rs.tokio.console.common.Id parent_resource_id = 6;
   */
  parentResourceId?: Id;

  /**
   * Is the resource an internal component of another resource?
   *
   * For example, a `tokio::time::Interval` resource might contain a 
   * `tokio::time::Sleep` resource internally.
   *
   * @generated from field: bool is_internal = 7;
   */
  isInternal = false;

  constructor(data?: PartialMessage<Resource>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.resources.Resource";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "id", kind: "message", T: Id },
    { no: 2, name: "metadata", kind: "message", T: MetaId },
    { no: 3, name: "concrete_type", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "kind", kind: "message", T: Resource_Kind },
    { no: 5, name: "location", kind: "message", T: Location },
    { no: 6, name: "parent_resource_id", kind: "message", T: Id },
    { no: 7, name: "is_internal", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Resource {
    return new Resource().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Resource {
    return new Resource().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Resource {
    return new Resource().fromJsonString(jsonString, options);
  }

  static equals(a: Resource | PlainMessage<Resource> | undefined, b: Resource | PlainMessage<Resource> | undefined): boolean {
    return proto3.util.equals(Resource, a, b);
  }
}

/**
 * The kind of resource (e.g. timer, mutex).
 *
 * @generated from message rs.tokio.console.resources.Resource.Kind
 */
export class Resource_Kind extends Message<Resource_Kind> {
  /**
   * Every resource is either a known kind or an other (unknown) kind.
   *
   * @generated from oneof rs.tokio.console.resources.Resource.Kind.kind
   */
  kind: {
    /**
     * `known` signals that this kind of resource is known to the console API.
     *
     * @generated from field: rs.tokio.console.resources.Resource.Kind.Known known = 1;
     */
    value: Resource_Kind_Known;
    case: "known";
  } | {
    /**
     * `other` signals that this kind of resource is unknown to the console API.
     *
     * @generated from field: string other = 2;
     */
    value: string;
    case: "other";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<Resource_Kind>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "rs.tokio.console.resources.Resource.Kind";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "known", kind: "enum", T: proto3.getEnumType(Resource_Kind_Known), oneof: "kind" },
    { no: 2, name: "other", kind: "scalar", T: 9 /* ScalarType.STRING */, oneof: "kind" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Resource_Kind {
    return new Resource_Kind().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Resource_Kind {
    return new Resource_Kind().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Resource_Kind {
    return new Resource_Kind().fromJsonString(jsonString, options);
  }

  static equals(a: Resource_Kind | PlainMessage<Resource_Kind> | undefined, b: Resource_Kind | PlainMessage<Resource_Kind> | undefined): boolean {
    return proto3.util.equals(Resource_Kind, a, b);
  }
}

/**
 * `Known` collects the kinds of resources that are known in this version of the API.
 *
 * @generated from enum rs.tokio.console.resources.Resource.Kind.Known
 */
export enum Resource_Kind_Known {
  /**
   * `TIMER` signals that this is a timer resource, e.g. waiting for a sleep to finish.
   *
   * @generated from enum value: TIMER = 0;
   */
  TIMER = 0,
}